Rexture more of the cockpit
Add canvas PFD. The SVG / Nasal only displays two parameters currently; but the menu selection and page framework is all wired in.

There will need to be multiple instances of this for each display. I think each display is capable of the same functions ; but not sure about that.

From: Richard Harrison <rjh@zaretto.com>