757-200: bug fix: make the engines visible over multiplayer
The engines were not visible over multiplayer because the 757-200.xml
model file selected on of two possible types of engines depending on
the value of the property /sim/engines, which is not visible over
multiplayer.  Consequently, the file would not select any engine model
when running in a remote instance of FlightGear.  The fix consists in
replacing /sim/engines with /sim/model/variant, an integer property
that is transmitted over the multiplayer protocol (see
http://wiki.flightgear.org/Multiplayer_protocol).