CRJ700-family: bug fix: do not prevent other planes' flaps from working over multiplayer.
Models/Instruments/PFD/pfd.xml refers to the property /sim/flaps/current-setting
(note the leading slash).  When a local aircraft loads a remote CRJ700 over multiplayer,
this causes the property to be created in the local aircraft's property tree.  This
in turns fools /usr/share/games/flightgear/Nasal/controls.nas into believing that
the local aircraft also has properties /sim/flaps/settings[0..N] to configure the
various flap steps.  If that is the case then the remote CRJ700 overrides the local
aircraft's flaps current setting; otherwise this causes [ and ] keys to stop working
in the local aircraft.

The fix consists in removing the leading slash: this causes the property to be created
in the property tree of the remote aircraft, as intended.