Add-ons: update addon-metadata.xml files following FG commit c3b1442546
Concerning the "GNU GPL version 2 or later" set for ATCChatter and its
authors: this is based on 1) the LICENSE file and 2) Torsten's message
on flightgear-devel from Thu, 07 Dec 2017
(<https://sourceforge.net/p/flightgear/mailman/message/36150299/>) which
said that the add-on was previously in FGData, and thus a priori
following the project licensing policy (GPLv2+ for FGData).

For SpokenATC and SpokenGCA, this was agreed by Rodolfo Leibner using
private messages on the FlightGear forum.