Citation II version 4.1.0: detailed modeling of the COM, NAV and ATC avionics devices, by Sascha Reißner.
CTL-22 (comm radio control):
This device has a volatile memory to store 6 frequencies.
If you are in the PRESELECT mode and press the STO (store) button,
the upper line shows [CH #] where '#' stays for a digit from 1 to 6.
This means 'Channel' (memory position) to store the frequency shown in
the lower line. The channel can be chosen with the toggle on the right
side pressing down (right click) to 'MEM'.
If you have the position that you wish, press the STO button again and
the frequency will be stored at that position.
If you press nothing, the device returns to normal operation after 5 seconds.
Whenever a frequency is shown in the lower line that is also
stored in memory, the MEM annunciator is lit (as a hint).
To call back the frequency from the memory, press the toggle to MEM.
Every press will advance to the next position (wrap around from 6 to 1)
and set the stored frequency in the preselect line.
If no frequency is stored at that position, nothing will happen.

Turning the device off clears its memory.

Pressing the ACT button switches the device into ACTIVE mode.
The lower line shows only dashes. Tuning the frequency changes the
active frequency.
Pressing the ACT button again switches the device back to PRESELECT
mode.

The power knob can be turned to 'SQ-OFF'.  This means that the squelch
is off, allowing you to hear noise even though no valid station is
tuned or the signal from the station is too weak.

The display shows a TX annunciator if you push the talk button for
FGCom (normaly 'space'). But FGCom always sets 'ptt' under comm[0],
even if it is switched to comm[1]. This is a bug in FGCom (as of
flightgear 2018.1.1 release candidate).

The display shows the ACT annunciator if the device is in ACTIVE mode
or if you swap the preset and active frequencies.  So ACT will allways
be lit, if the device send a frequency to the radio itself.

CTL-32 (nav radio control):
The same memory behavior as the CTL-22.
Has also a ACTIVE mode like the CTL-22.
The power knob can be turned to HOLD, but I don't know what the real
device does in this mode.  So, HOLD is currently the same as ON.

CTL-62 (ADF radio control):
I have read that this device has only 4 memory positions, but i have
left it at 6. So, the memory behaves as the CTL-22.
Has also a ACTIVE mode like the CTL-22.
The power knob has the modes ANT, ADF and TONE.

ANT (for Antenna) uses only the sense antenna and will be used from
the pilot to verify that he/she has the right NDB (listen the morse
code).  In this position, the ADF can't do anything (no bearing).

ADF uses both antennae (sense and loop) to find the bearing to the
station. In this mode, if you turn up the audio, nothing should be
heard, because the ADF searches for the weakest signal.

TONE will be used on some old NDBs. The morse code will be sent in
another fashion. If you don't hear anything in ANT, try TONE.  TONE on
some other devices is known as 'BFO' (Beat Frequency Oscillator). As
far as I know, the ADF cannot find the bearing to the station in BFO
mode.

CTL-92 (transponder control):
This device can control 1 or 2 transponders.
If there is only one transponder, the device has no switch on the
right side. The hole will be closed with a black plastic cap.
All changes to the squawk code goes directly to the transponder.

If there are 2 transponders, changing the squawk will apply to the
inactive transponder. To make it active, toggle the switch on the
right side.

The IDENT button works correctly and if a ident is running, the ACT
annunciatior is lit on the display.

The models of the devices also include the TEST button and display
'diAG' for testing but I don't know how it works exactly.
There is also a 'RMT' annunciator in the display. I think it stands for
'Remote', but I don't know when they should lit.
For further developing these devices, I need a manual or a pilot who
knows them.

This update includes also a correction for the bearing (red) needle of
the HSIs. Without a signal, the needle was not exactly to the right.