Ryan Navion / Super Navion 260 [0.4.6] --------------------------------------
added:
- nav[1] for KNS80
- dme[1] for KNS80
- nav-source to select the navigation source (under 'instrumentation/nav-source')

changed:
- KNS80 model little bit updated
- KNS80 calculates RNAV with property rules (nasal deacitvated)
    (i'm sure there are some bugs)
- KAP140 updated to use the selected nav-source
- HSI (KI252a) updated to use the selected nav-source
- parking brake set at start up

info:
The AP (KAP140) use allways the values that are selected in 'nav-source'.
Per default, 'instrumentation/nav-source/selector' is set to '0' for the nav radio (KX165).
If you want to test the KNS80, set 'instrumentation/nav-source/selector' to '1'.
It is also prepared for a third source (like GPS).
To accomplish that, configure the values in 'nav-source-config.xml' properly and
set 'instrumentation/nav-source/selector' to '2' for the third navigation source.